+++ src/modules/ESP8266.h
@@ -11,17 +11,17 @@ class ESP8266: public TransportLayer {
     ESP8266(Module* module);
     
     // basic methods
-    uint8_t begin(long speed);
-    uint8_t reset();
-    uint8_t join(const char* ssid, const char* password);
+    int16_t begin(long speed);
+    int16_t reset();
+    int16_t join(const char* ssid, const char* password);
     
     // transport layer methods (implementations of purely virtual methods in TransportMethod class)
-    uint8_t openTransportConnection(const char* host, const char* protocol, uint16_t port, uint16_t tcpKeepAlive = 0);
-    uint8_t closeTransportConnection();
-    uint8_t send(const char* data);
-    uint8_t send(uint8_t* data, uint32_t len);
+    int16_t openTransportConnection(const char* host, const char* protocol, uint16_t port, uint16_t tcpKeepAlive = 0);
+    int16_t closeTransportConnection();
+    int16_t send(const char* data);
+    int16_t send(uint8_t* data, uint32_t len);
     size_t receive(uint8_t* data, size_t len, uint32_t timeout = 10000);
-    uint16_t getNumBytes(uint32_t timeout = 10000, size_t minBytes = 10);
+    size_t getNumBytes(uint32_t timeout = 10000, size_t minBytes = 10);
     
   private:
     Module* _mod;
